MAIN PAGE WORKS
THE MAIN PAGE WORKS! Now I need to work on the edit functionality. The
links don't quite work yet. The edit.php page is broken. Also made a
slight change to settings where if you don't have a class that period,
then leave it blank, and everything else will revolve around whether or
not the subtype is empty.