Attempting to do file uploading for edit too. Problems
Okay, so you can't retrieve the file from the server. I think the thing
we need to do is, when the file is uploaded, we need to change the name
too something generic, like an id, 00001, or something. We also need to
make another table for file locations with an id that goes into the
announcements column that references this file locations table. When an
announcement is edited, the newly uploaded file will have it's own name
and id in the file locations table, and this id will be substituted for
the one currently in the announcements.